Enum Consistency in ConfigSimulationDialog
Switch-case-based conversion methods between enums and integers have been removed in favour of directly using enums in both cases.
This should stop any discrpency occuring if all enums refer to a central source.

This could not be applied to Optimisation.cpp however during loading and saving parameter methods as this could cause old inputs not to work correctly. This could be changed in the future.