ParallelCoordinatesPlotter First Version
This commit is for reference only and is not stable.

Parallel coordinates plotter class created.
Current version can be used to plot a parallel coordinates graph of a set of boundary points.
Is suitable to be used when displaying a finishied optimisaiton, however starting a new optimisation with this in use may cause the program to crash.