Optimisations now store the initial mesh boundary
The initial mesh boundary is saved into a new file called 'Boundary_Points.txt'
This is different to the initial profile as the mesh is an approximation of the initial profile.

Read / Write methods have been added in the Optimisation class.
Much of it has been moved out of Optimisation::readProfilePointsFromSimulationDir() as this was incomplete and it was not possible to add the additional Boundary Point info to another input file.

Supplementary methods added in MeshDialog and BoundaryPointModel classes to fetch BoundaryPoints.

AerOpt Settings updated to be aware of new file.