PC Graph included in program execution
PC Graph window will be opened when a new optimisation index is set in MainWindow.

Please note: Unless the optimisation has finished executing then the graph displayed will be blank.
This includes if the optimisation was stopped before the end.

This is temporary and will be fixed in a future commit.