Simplified Fitness Plot Touch Functionality
Pinch-to-zoom working for fitness graph by making additions to qcustomplot

Note: The way pinch-to-zoom works here is not consistent with how it works for other parts of the UI due to QCustomPlot's differing behaviour and complexity of implementation.

To pinch-to-zoom, hold one finger still whilst moving the other finger to pinch.
Otherwise, the pinch gets mistaken for a pan and hte behaviour doesn't work as well.

Ideally this should be fixed, but may not be worth the time investment given that the functionality works to some degree.